improvement(heroku): correct ormconfig creation
