+++ api/src/database/database.providers.ts
@@ -1,43 +1,6 @@
 import { createConnection } from 'typeorm';
 import { DATABASE_CONNECTION } from 'src/constants';
-
-import { parse } from 'pg-connection-string';
-
-let config;
-// On heroku we use a connection string
-if (process.env.DATABASE_URL) {
-  const databaseUrl = process.env.DATABASE_URL;
-  const connectionOptions = parse(databaseUrl);
-  config = {
-    type: 'postgres',
-    host: connectionOptions.host,
-    port: parseInt(connectionOptions.port, 10),
-    username: connectionOptions.user,
-    password: connectionOptions.password,
-    database: connectionOptions.database,
-    synchronize: false,
-    entities: [
-      __dirname + '/../**/*.entity{.ts,.js}',
-    ],
-    extra: {
-      ssl: connectionOptions.ssl,
-    },
-  };
-} else {
-  config = {
-    type: 'postgres',
-    host: process.env.TYPEORM_HOST,
-    port: parseInt(process.env.TYPEORM_PORT),
-    username: process.env.TYPEORM_USERNAME,
-    password: process.env.TYPEORM_PASSWORD,
-    database: process.env.TYPEORM_DATABASE,
-    entities: [
-      __dirname + '/../**/*.entity{.ts,.js}',
-    ],
-    synchronize: false,
-  };
-}
-// end dupe
+import config from './ormconfig';
 
 const connectionPromise = createConnection(config);
